Move workspaceTheme to src/lib, so it is included in the generated lib directory
